removed unneeded code, fixed env
