+++ beta-src/.gitignore
@@ -8,9 +8,9 @@
 
 # misc
 .DS_Store
+.env
 .env.local
-.env.development.local
-.env.test.local
-.env.production.local
+.env.production
+.env.test
 
 npm-debug.log*
