+++ beta-src/package.json
@@ -27,6 +27,8 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build && rm -rf ../beta && mv build ../beta",
+    "build-local": "NODE_ENV=local react-scripts build --local --configuration=local && rm -rf ../beta && mv build ../beta",
+    "build-test": "NODE_ENV=test react-scripts build --test --configuration=test && rm -rf ../beta && mv build ../beta",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "svgr": "svgr -d src/components/svgr-components --typescript --ignore-existing src/assets/svg-svgr"
