+++ beta-src/src/enums/ApiRoute.ts
@@ -1,0 +1,11 @@
+enum ApiRoute {
+  GAME_TOGGLEVOTE = "game/togglevote",
+  GAME_STATUS = "game/status",
+  GAME_OVERVIEW = "game/overview",
+  GAME_ORDERS = "game/orders",
+  GAME_SENDMESSAGE = "game/sendmessage",
+  PLAYERS_CD = "players/cd",
+  PLAYERS_MISSING_ORDERS = "players/missing_orders",
+}
+
+export default ApiRoute;
